pyproject.toml
[project]
name = "py-esc-pos"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.1",
    "flask-cors>=6.0.0",
    "pillow>=11.2.1",
    "pyinstaller>=6.13.0",
    "pymupdf>=1.26.0",
    "python-escpos>=3.1",
    "pyusb>=1.3.1",
    "werkzeug>=3.1.3",
]
